refactor: adjust the directory
