+++ src/widgets/antd/map.jsx
@@ -1,3 +1,2 @@
 import map from '../common/map';
-
 export default map;
