+++ src/widgets/fusion/map.jsx
@@ -1,3 +1,2 @@
-import map from '../common/map';
-
+import map from '../../components/map';
 export default map;
